+++ res/history.md
@@ -6,4 +6,4 @@
 
 ## Migrated from Total Replay
 
- - ...
+ - Summer Games